% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daoh-package.R
\docType{package}
\name{daoh-package}
\alias{daoh}
\alias{daoh-package}
\title{daoh: Days Alive and Out of Hospital (DAOH) Calculation}
\description{
Calculates Days Alive and Out of Hospital (DAOH) from administrative
admission/discharge/mortality data using three algorithms (nights, days,
exact) and three death-handling approaches (midday, midnight, zero).
Includes tools for comparing methods via Bland-Altman analysis, intraclass
correlation, and reclassification statistics, plus plotting functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davecumin/daoh}
  \item Report bugs at \url{https://github.com/davecumin/daoh/issues}
}

}
\author{
\strong{Maintainer}: David Cumin \email{d.cumin@auckland.ac.nz}

Authors:
\itemize{
  \item David Cumin \email{d.cumin@auckland.ac.nz}
}

}
\keyword{internal}
